requirements.txt
pymysql==1.1.0
mariadb==1.1.6
sqlalchemy==2.0.28
alembic==1.12.1
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0
matplotlib==3.8.3
seaborn==0.13.1
plotly==5.19.0
flask==3.0.2
flask-restful==0.3.10
flask-cors==4.0.0
dash==2.15.0
dash-bootstrap-components==1.5.0
dash-ag-grid==2.4.0
cryptography==42.0.5
statsmodels==0.14.1
openpyxl==3.1.2
jupyter==1.0.0
python-dotenv==1.0.1
apscheduler==3.10.4
redis==5.0.1
gunicorn==21.2.0
